solved error for numbers text

+++ get_next_line.h
@@ -1,15 +1,21 @@
-//
-//  get_next_line.h
-//  get_next_line
-//
-//  Created by Faruk Kuz on 7.08.2025.
-//
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   get_next_line.h                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: fakuz <fakuz@student.42istanbul.com.tr>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/08/27 10:37:23 by fakuz             #+#    #+#             */
+/*   Updated: 2025/08/30 20:20:04 by fakuz            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #ifndef GET_NEXT_LINE_H
 # define GET_NEXT_LINE_H
 
 # include <stdlib.h>
 # include <unistd.h>
+# include <stdio.h>
 
 # ifndef BUFFER_SIZE
 #  define BUFFER_SIZE 42
@@ -17,7 +23,7 @@
 
 char	*get_next_line(int fd);
 size_t	ft_strlen(char *s);
-int		ft_strchr(char *s, int c);
+size_t	ft_strchr(char *s, int c);
 char	*ft_strdup(char *s);
 char	*ft_substr(char *s, unsigned int start, size_t len);
 char	*ft_strjoin(char *s1, char *s2);
